Changed new list segment to take camp resource by value, filled out remaining index set tests
