+++ test/functional/forall/tests/test-forall-basic-indexset.hpp
@@ -5,46 +5,56 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-#ifndef __TEST_FORALL_LISTSEGMENT_HPP__
-#define __TEST_FORALL_LISTSEGMENT_HPP__
+#ifndef __TEST_FORALL_BASIC_INDEXSET_HPP__
+#define __TEST_FORALL_BASIC_INDEXSET_HPP__
 
-#include "test-forall-segment.hpp"
+#include "test-forall-indexset-utils.hpp"
+
 #include <cstdio>
-#include <cstdlib>
-#include <cstring>
-#include <ctime>
 #include <algorithm>
-#include <numeric>
+#include <random>
+#include <vector>
 
 using namespace camp::resources;
 using namespace camp;
 
+
 template <typename INDEX_TYPE, typename WORKING_RES, typename EXEC_POLICY>
-void ForallListSegmentTest(INDEX_TYPE N)
+void ForallISetTest()
 {
 
-  // Create and initialize indices in idx_array used to create list segment
-  std::vector<INDEX_TYPE> idx_array;
+  using RangeSegType       = RAJA::TypedRangeSegment<INDEX_TYPE>;
+  using RangeStrideSegType = RAJA::TypedRangeStrideSegment<INDEX_TYPE>;
+  // RDH   using ListSegType        = RAJA::TypedListSegment<INDEX_TYPE>;
+  using ListSegType        = RAJA::NewTypedListSegment<INDEX_TYPE>;
 
-  srand ( time(NULL) );
-
-  for (INDEX_TYPE i = 0; i < N; ++i) {
-    INDEX_TYPE randval = rand() % N;
-    if ( i < randval ) {
-      idx_array.push_back(i);
-    }     
-  }
-
-  size_t idxlen = idx_array.size();
+  using IndexSetType = 
+   RAJA::TypedIndexSet< RangeSegType, RangeStrideSegType, ListSegType >; 
 
   Resource working_res{WORKING_RES()};
 
-  // Create list segment for tests
-  RAJA::NewTypedListSegment<INDEX_TYPE> lseg(&idx_array[0], idxlen,
-                                             working_res);
-//RAJA::NewTypedListSegment<INDEX_TYPE> lseg(&idx_array[0], idxlen);
-//RAJA::TypedListSegment<INDEX_TYPE> lseg(&idx_array[0], idxlen);
+  INDEX_TYPE last_idx = 0;
 
+  IndexSetType iset; 
+  buildIndexSet<INDEX_TYPE, RangeSegType, RangeStrideSegType, ListSegType>(
+    iset, last_idx, working_res);
+
+
+  //
+  // Collect actual indices in index set for testing.
+  //
+  std::vector<INDEX_TYPE> is_indices; 
+  getIndices(is_indices, iset);
+
+ 
+  //
+  // Working array length
+  //
+  const INDEX_TYPE N = last_idx + 1;
+
+  //
+  // Allocate and initialize arrays used in testing
+  //  
   INDEX_TYPE* working_array;
   INDEX_TYPE* check_array;
   INDEX_TYPE* test_array;
@@ -59,11 +69,13 @@ void ForallListSegmentTest(INDEX_TYPE N)
 
   working_res.memcpy(working_array, test_array, sizeof(INDEX_TYPE) * N);
 
-  std::for_each( std::begin(idx_array), std::end(idx_array), 
-                 [=](INDEX_TYPE& idx ) { test_array[idx] = idx; }
+  std::for_each( std::begin(is_indices), std::end(is_indices), 
+                 [=](INDEX_TYPE& idx) { 
+                   test_array[idx] = idx; 
+                 }
                );
 
-  RAJA::forall<EXEC_POLICY>(lseg, [=] RAJA_HOST_DEVICE(INDEX_TYPE idx) {
+  RAJA::forall<EXEC_POLICY>(iset, [=] RAJA_HOST_DEVICE(INDEX_TYPE idx) {
     working_array[idx] = idx;
   });
 
@@ -81,18 +93,13 @@ void ForallListSegmentTest(INDEX_TYPE N)
 }
 
 
-TYPED_TEST_P(ForallSegmentTest, ListSegmentForall)
+TYPED_TEST_P(ForallIndexSetTest, IndexSetForall)
 {
   using INDEX_TYPE       = typename at<TypeParam, num<0>>::type;
   using WORKING_RESOURCE = typename at<TypeParam, num<1>>::type;
   using EXEC_POLICY      = typename at<TypeParam, num<2>>::type;
 
-  ForallListSegmentTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>(13);
-
-  ForallListSegmentTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>(2047);
-
-  ForallListSegmentTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>(32000);
-
+  ForallISetTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>();
 }
 
-#endif  // __TEST_FORALL_LISTSEGMENT_HPP__
+#endif  // __TEST_FORALL_BASIC_INDEXSET_HPP__
